requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-deepseek>=0.1.4
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
deepseek-ai>=0.0.1 